Fixed major pointer bugs, and misc bugs
Fixed emails, and keyword checkboxes, so when we delete either a user can not start scrape, or upload a file until finish. Also a user will be unable to delete keywords, or emails unless the thread emits a signal saying its ok.
Fixed curl single, and multi config. Added a spinbox option so user can choose whether to use multi or a single url for harvesting